refactor: centralize API error rendering
