+++ lmdeploy/serve/openai/endpoints/completions.py
@@ -14,6 +14,11 @@ from fastapi.responses import StreamingResponse
 from lmdeploy.pytorch.disagg.conn.protocol import MigrationRequest
 from lmdeploy.serve.core.exceptions import ErrorCode, RequestError
 from lmdeploy.serve.openai.endpoints.common import build_serving_generation_config, validate_request
+from lmdeploy.serve.openai.errors import (
+    create_error_response,
+    create_request_error_response,
+    request_error_payload,
+)
 from lmdeploy.serve.openai.protocol import (
     CompletionRequest,
     CompletionResponse,
@@ -22,11 +27,6 @@ from lmdeploy.serve.openai.protocol import (
     CompletionStreamResponse,
     LogProbs,
     UsageInfo,
-)
-from lmdeploy.serve.openai.utils import (
-    create_error_response,
-    create_request_error_response,
-    request_error_payload,
 )
 from lmdeploy.serve.utils.request_cleanup import with_request_cleanup
 from lmdeploy.serve.utils.server_utils import validate_json_request
